refactor: create module directory for cache
